+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4784,7 +4784,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KADY1CTE6KJ8WM96HN1Z09Z6"),
+  id: z.string().default("01KAE3ET2RDP05HJ0J0VETNQ4H"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -4803,7 +4803,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KADY1CTE6KJ8WM96HN1Z09Z6"),
+  id: z.string().default("01KAE3ET2RDP05HJ0J0VETNQ4H"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),