+++ packages/orq-rc/RELEASES.md
@@ -5159,3 +5159,13 @@ Based on:
 - [typescript v4.0.0-rc.50] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.50 - packages/orq-rc
+
+## 2025-11-19 13:05:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.51] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.51 - packages/orq-rc